chore: moved githubManager from web/lib to lib/services
